Removes redundant data, makes the template smarter
We now put everything the Javascript needs somewhere in the HTML upon
site generation. About all we really need is the name of the project's
GitHub repo. On that note, some of the repos (namely, api.data.gov and
18f.gsa.gov) have dots in the name and can't be used as they are for a css
ID. In this case we have a special YAML field called css_id we can use to
override the default behavior. Pretty cool! Check it out.