Update with more recent metrics from https://api.data.gov/metrics/
In our last round of metric updates, we also added a "since" date to the
numbers mean a little more, so I made that same change here. Let me know
if anyone feels differently.

(and we have plans to API-ify these metrics, so at some point, we might
have a better way to maintain these).