Adds redirect plugin, modifies dashboard plugin
The Dashboard plugin now lowercases all 'name' values for projects and
automatically adds the redirect for the original uppercased string.

For example, the project betaFEC has the name `betaFEC`. Before this
change that would result in a url `dashboard/project/betaFEC/`. After
thi change it will result in a url `dashboard/project/betafec` that also
responds to the named url.

Lowercase urls are a specific rule in the 18F content guide. This commit
also ensures that URLs listed on the index page are written lower case.