Display preview based on URL parameter
We have been discussing a live preview for quite some time, so this is
a first implementation attempt at it. Essentially, you can pass in a
`url` parameter in the query string that points to an `.about.yml`
anywhere on the internet.

It then parses the YML and inserts it into the correct spots in the DOM.