Use yamlJS as browser compatible YAML parser
I wanted to test another yaml parser because `js-yaml` does not work
well with browserify. `yamljs` works just fine. I updated all
references to it, and tried to override the global jsYaml object so
that you can easily comment out that dependency within the source code
of the validator module and then the validator can consume this module.

This is a stopgap. We need the validator module to make this change
upstream.