Updates import public and dashboard plugin
Import public now fetches all project data from the Team API and then
dumps it into a file at `_data/all.json`. From there we generate a YAML
file for each project we intend to use in the Dashboard and generate the
pages from those data files.

The advantage of this is we can experiment with the structure of the
data we want to use for the Dashboard without committing the Team API
team to implementing anything and we can also modify anything that comes
through incorrectly from the Team API and get rid of fields we don't
want like `errors` after we inspect them and attempt to address them
wherever appropriate.