Adding support for subprojects and repo links.
Subprojects:
Subprojects are designated with a `parent` tag in the about.yml. The dashboard plugin will pull them into the parent repo in a children key.

Repo Links:
Repo Links are designated by setting the link `category` to `repo`. The link text should be either a full url to a github repo (https://github.com/18F/federalist) or a org/name slug (18F/federalist). Repo links will not show up in the related links, but will be listed in the `code` section of the site. A repo link has precedence over a subproject; a subproject will not be listed as such if it is also included as a repo link.

Layout/Include Changes:
- The `code` and `links` sections were pulled into seperate includes.
- A new style was added to for `h2` project name headers.
- The code to display a repo's license information was simplified.

Concerns:
- The repo links functionality and subprojects won't 'just work' with the ongoing `preview` functionality that is being developed, since both of those pieces are done in the dashboard plugin.
- Not thrilled with the layout for multiprojects, I'd love to get some suggestions.