Adds support for either blog or blogTag entries.
Previously, about.yml used `blog` to define the blog tags to be called out for the individual project. The tag is in the process of being changed to `blogTag`. This commit will use either, but prefer `blogTag` if it is present.

Additionally, added a `strip` for the list of tags, because the common mistake of putting spaces between tags (i.e. `projectname, opendata, api` would yield incorrect urls for `opendata` (`https://18f.gsa.gov/tags/ opendata`) and `api`.

Resolves #310